+++ BookStoreMVC.Utility/Constants.cs
@@ -4,8 +4,8 @@
     {
         public static class Prices
         {
-            public const double Shipping = 5.0;
-            public const double Vat = 0.23;
+            public const decimal Shipping = 5.0m;
+            public const decimal Vat = 0.23m;
         }
     }
 }
